fix 2

+++ database/migrations/2025_06_26_175605_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('price', 10, 2);
+            $table->double('price');
             $table->integer('stock');
             $table->string('image')->nullable();
             $table->timestamps();
